+++ ironflow/nodes/std/special_nodes.py
@@ -3,7 +3,7 @@ from contextlib import redirect_stdout, redirect_stderr
 
 from ironflow.NENV import *
 widgets = import_widgets(__file__)
-from ironflow.ironflow.panels.node_interface.input_widgets import SliderControl
+from ironflow.ironflow.boxes.node_interface.input_widgets import SliderControl
 
 
 class NodeBase(Node):
